Initial replacement of old `Biquad` with `BlockBiquad`
