+++ shared-module/synthio/Biquad.h
@@ -1,3 +1,4 @@
+
 // This file is part of the CircuitPython project: https://circuitpython.org
 //
 // SPDX-FileCopyrightText: Copyright (c) 2023 Jeff Epler for Adafruit Industries
@@ -6,15 +7,23 @@
 
 #pragma once
 
-#include "py/obj.h"
+#include "shared-bindings/synthio/Biquad.h"
+#include "shared-module/synthio/block.h"
 
 #define BIQUAD_SHIFT (15)
 
-typedef struct {
+typedef struct synthio_biquad {
+    mp_obj_base_t base;
+    synthio_filter_mode mode;
+    synthio_block_slot_t f0, Q, A;
+    mp_float_t cached_W0, cached_Q, cached_A;
     int32_t a1, a2, b0, b1, b2;
+} synthio_biquad_t;
+
+typedef struct {
     int32_t x[2], y[2];
 } biquad_filter_state;
 
-void synthio_biquad_filter_assign(biquad_filter_state *st, mp_obj_t biquad_obj);
+void common_hal_synthio_biquad_tick(mp_obj_t self_in);
 void synthio_biquad_filter_reset(biquad_filter_state *st);
-void synthio_biquad_filter_samples(biquad_filter_state *st, int32_t *buffer, size_t n_samples);
+void synthio_biquad_filter_samples(mp_obj_t self_in, biquad_filter_state *st, int32_t *buffer, size_t n_samples);
