+++ shared-module/synthio/__init__.h
@@ -21,7 +21,6 @@
 #include "shared-module/audiocore/__init__.h"
 #include "shared-bindings/synthio/__init__.h"
 #include "shared-bindings/synthio/Biquad.h"
-#include "shared-bindings/synthio/BlockBiquad.h"
 
 typedef struct {
     uint16_t dur;
@@ -96,8 +95,3 @@ int synthio_sweep_in_step(synthio_lfo_state_t *state, uint16_t dur);
 extern mp_float_t synthio_global_rate_scale, synthio_global_W_scale;
 extern uint8_t synthio_global_tick;
 void shared_bindings_synthio_lfo_tick(uint32_t sample_rate, uint16_t num_samples);
-
-static inline bool synthio_is_any_biquad(mp_obj_t biquad_maybe) {
-    return mp_obj_is_type(biquad_maybe, &synthio_block_biquad_type_obj)
-           || mp_obj_is_type(biquad_maybe, (const mp_obj_type_t *)&synthio_biquad_type_obj);
-}
