+++ tests/circuitpython/audiofilter_filter_biquads.py
@@ -1,14 +1,14 @@
 from audiofilters import Filter
 from audiofilterhelper import synth_test, white8k
-from synthio import BlockBiquad, FilterMode
+from synthio import Biquad, FilterMode
 
 
 @synth_test
 def basic_filter():
     effect = Filter(
         filter=[
-            BlockBiquad(FilterMode.LOW_PASS, 400),
-            BlockBiquad(FilterMode.HIGH_PASS, 300, Q=8),
+            Biquad(FilterMode.LOW_PASS, 400),
+            Biquad(FilterMode.HIGH_PASS, 300, Q=8),
         ],
         bits_per_sample=16,
         samples_signed=True,
